+++ led_strip/interface/led_strip_interface.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2022 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2022-2025 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -13,6 +13,8 @@ extern "C" {
 #endif
 
 typedef struct led_strip_t led_strip_t; /*!< Type of LED strip */
+
+typedef struct led_strip_group_t led_strip_group_t; /*!< Type of LED group strip */
 
 /**
  * @brief LED strip interface definition
@@ -67,6 +69,30 @@ struct led_strip_t {
     esp_err_t (*refresh)(led_strip_t *strip);
 
     /**
+     * @brief Refresh memory colors to LEDs asynchronously
+     *
+     * @param strip: LED strip
+     *
+     * @return
+     *      - ESP_OK: Refresh successfully
+     *      - ESP_FAIL: Refresh failed because some other error occurred
+     *
+     * @note:
+     *      This function is non-blocking, so you need to call `led_strip_refresh_wait_async_done` to wait for the refresh to complete before modifying the LED colors again.
+     */
+    esp_err_t (*refresh_async)(led_strip_t *strip);
+
+    /**
+     * @brief Wait for the async refresh to complete
+     *
+     * @param strip: LED strip
+     *
+     * @return
+     *      - ESP_OK: Wait for the async refresh to complete successfully
+     */
+    esp_err_t (*refresh_wait_async_done)(led_strip_t *strip);
+
+    /**
      * @brief Clear LED strip (turn off all LEDs)
      *
      * @param strip: LED strip
@@ -88,6 +114,35 @@ struct led_strip_t {
      *      - ESP_FAIL: Free resources failed because error occurred
      */
     esp_err_t (*del)(led_strip_t *strip);
+};
+
+
+/**
+ * @brief LED strip group interface definition
+ */
+struct led_strip_group_t {
+    /**
+     * @brief Get LED strip handle by index
+     *
+     * @param group: LED strip group
+     * @param index: LED strip index
+     * @param ret_strip: Retured LED strip handle
+     *
+     * @return
+     *     - ESP_OK: Success
+     *     - ESP_ERR_INVALID_ARG: Invalid argument
+     */
+    esp_err_t (*get_strip_handle)(led_strip_group_t *group, uint8_t index, led_strip_handle_t *ret_strip);
+    /**
+     * @brief Free LED strip group resources
+     *
+     * @param group: LED strip group
+     *
+     * @return
+     *      - ESP_OK: Free resources successfully
+     *      - ESP_FAIL: Free resources failed because error occurred
+     */
+    esp_err_t (*del)(led_strip_group_t *group);
 };
 
 #ifdef __cplusplus